move very slow test to it's own module
